error handling added

+++ TicketManagement.Application/Features/Events/Commands/DeleteEvent/DeleteEventCommandHandler.cs
@@ -6,6 +6,8 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using TicketManagement.Application.Contracts.Persistence;
+using TicketManagement.Application.Exceptions;
+using TicketManagement.Domain.Entities;
 
 namespace TicketManagement.Application.Features.Events.Commands.DeleteEvent
 {
@@ -23,6 +25,11 @@ namespace TicketManagement.Application.Features.Events.Commands.DeleteEvent
         public async Task<Unit> Handle(DeleteEventCommand request, CancellationToken cancellationToken)
         {
             var eventToDelete = await _eventRepository.GetByIdAsync(request.EventId);
+
+            if (eventToDelete == null)
+            {
+                throw new NotFoundException(nameof(Event), request.EventId);
+            }
 
             await _eventRepository.DeleteAsync(eventToDelete);
